+++ scratch.js
@@ -1,7 +1,15 @@
 var nlp = require('./src/index');
-nlp.verbose('tagger');
+// nlp.verbose('tagger');
 
 // var doc = nlp('i just walked to the store');
 // doc.match('just [#PastTense] to the store').debug();
 
-nlp('slowly').debug(); //.adverbs().debug();
+// nlp('cheering hard').debug();
+
+var nlp2 = nlp.clone();
+nlp.addWords({
+  bat: 'Verb'
+});
+
+nlp('bat').debug();
+nlp2('bat').debug();