+++ scripts/browserTest.js
@@ -1,11 +1,11 @@
-var exec = require('shelljs').exec
+var exec = require('shelljs').exec;
 var browserifyGlob = '"./node_modules/.bin/browserify-glob"';
 var fileServer = '"./node_modules/.bin/http-server"';
 
 //run tests on the client-side
 var out = './test/client/compiled_tests.js';
 var cmd = browserifyGlob + ' "./test/unit/**/*.test.js" ';
-cmd += ' -t [ babelify --presets [ es2015 stage-2 ] ] > ' + out;
+cmd += ' -t [ babelify --presets [ es2015 ]] > ' + out;
 cmd += ' && ' + fileServer + ' test/client -o -c-1';
 exec(cmd);